everything moved should work
