+++ platform-operator/Makefile
@@ -128,10 +128,10 @@ docker-clean:
 	rm -rf ${DIST_DIR}
 
 .PHONY: docker-build
-docker-build: generate-bom go-build-linux docker-build-common
+docker-build: generate-bom generate-catalog go-build-linux docker-build-common
 
 .PHONY: docker-build-debug
-docker-build-debug: generate-bom go-build-linux-debug docker-build-common
+docker-build-debug: generate-bom generate-catalog go-build-linux-debug docker-build-common
 
 .PHONY: docker-build-common
 docker-build-common:
@@ -225,3 +225,8 @@ generate-bom:
 	mkdir out || true
 	../tools/scripts/generate_bom.sh verrazzano-bom.json ${VERRAZZANO_VERSION} ${VERRAZZANO_APPLICATION_OPERATOR_IMAGE} ${VERRAZZANO_CLUSTER_OPERATOR_IMAGE} ${VERRAZZANO_AUTHPROXY_IMAGE} ${DOCKER_IMAGE_NAME} ${DOCKER_IMAGE_TAG} out/generated-verrazzano-bom.json
 
+.PHONY: generate-catalog
+generate-catalog:
+	@echo Generating catalog with verrazzano version ${VERRAZZANO_VERSION}
+	mkdir out || true
+	../tools/scripts/generate_catalog.sh manifests/catalog/catalog.yaml ${VERRAZZANO_VERSION} out/generated-catalog.yaml
